Time: 3 ms (17.15%), Space: 7.5 MB (8.37%) - LeetHub

+++ 1291-sequential-digits/1291-sequential-digits.cpp
@@ -6,7 +6,7 @@ public:
         
         for(int i=0;i<9;i++){
             for(int j=i;j<9;j++){
-                long long num = stoi(s.substr(i,j-i+1));
+                int num = stoi(s.substr(i,j-i+1));
                 if(num>=low && num<=high)
                     ans.push_back(num);
                 else if(num>high)
